ref: organize raw dataset files to separate folder
